refactored some UAuth classes
