+++ src/main/java/app/Main.java
@@ -7,9 +7,9 @@ import entities.Quiz;
 import entities.SinglePlayer;
 import entities.Users.CommonUserFactory;
 import entities.Users.User;
-import interface_adapter.UAuthController;
-import interface_adapter.UAuthPresenter;
-import interface_adapter.UAuthViewModel;
+import interface_adapter.UAuth.UAuthController;
+import interface_adapter.UAuth.UAuthPresenter;
+import interface_adapter.UAuth.UAuthViewModel;
 import interface_adapter.ViewManagerModel;
 import interface_adapter.skip_song.SkipController;
 import interface_adapter.skip_song.SkipPresenter;
