+++ src/main/java/interface_adapter/UAuth/UAuthPresenter.java
@@ -1,6 +1,5 @@
-package interface_adapter;
+package interface_adapter.UAuth;
 
-import interface_adapter.UAuthViewModel;
 import interface_adapter.play_song.PlayViewModel;
 import usecase.UserAuth.UAuthOutputBoundary;
 import usecase.UserAuth.UAuthOutputData;
