+++ src/main/java/usecase/score/ScoreInputBoundary.java
@@ -1,4 +1,5 @@
 package usecase.score;
 
 public interface ScoreInputBoundary {
+    public int returnScore();
 }
