+++ src/main/java/usecase/score/ScoreInteractor.java
@@ -1,4 +1,8 @@
 package usecase.score;
 
-public class ScoreInteractor {
+public class ScoreInteractor implements ScoreInputBoundary{
+    @Override
+    public int returnScore() {
+        return 0;
+    }
 }
